(build) suppressed some warning messages
swept build permutations with both clang and g++, passed.

FossilOrigin-Name: 81e672f80a153c340eb1d95c72995f28b1f226229266506108a8487339e08045

+++ src/frameserver/encode/default/encode_presets.c
@@ -117,7 +117,7 @@ static bool default_acodec_setup(struct codec_ent* dst, unsigned channels,
 
 	LOG("(encode) audio encoder setup @ %d hz, requested quality: %d, "
 		"got %d kbit/s using %s\n", samplerate, abr,
-		ctx->bit_rate / 1000, codec->name);
+		(int)(ctx->bit_rate / 1000), codec->name);
 
 	if (avcodec_open2(dst->storage.audio.context,
 		dst->storage.audio.codec, NULL) != 0){
